refactor: Init the current node from endpoints.py
In some cases the endpoints refer to the bootstrap node.
In other cases the endpoints refer to the other nodes of the network.
The methods of the endpoints need the nodes as input.
To simplify the system, we initiate the node object of each session
from the endpoints.py file.