feat: Transfer the blockchain between the nodes
Add endpoints "/ask_chain" and "/post_chain".

Everytime the blockchain is updated with a verified block,
we need to send the new instance of the verified blockchain
to all the nodes in the network.
Each node should ask the last block's validator
for their version of the blockchain (endpoint = "/ask_chain")
and get their self.chain updated
with the last version of the blockchain (endpoint = "/post_chain")
before proceeding to the next block.