chore: Add runner script
For easier testing and running of the main code.
The runner script opens a separate Git Bash terminal
for each node in the network to run on.
For each node it executes "python src/app.py"
with all the required command-line arguments.

To use it you need to specify the path to git-bash.exe,
as well as the node ports.