refactor: Readjust transaction nonce attribute
The transaction nonce is not calculated until
all nodes in the network have validated the transaction,
and the transaction is ready to be added into the sender's wallet.

The validation of the transaction
includes the verification of its signature.

The nonce changes after the transaction has been signed
and thus it cannot be included into the attributes
upon which the transaction hash is calculated.