feat: Make a new transaction
To make a new transaction
make a "/make_transaction" POST request to the sender node.
Required Input: recipient_id, bcc, message

The "/make_transaction" POST request calls
the create_transaction node method on the sender node.

This method creates the transaction
if the sender node has sufficient balance.
It gets its signature and it broadcasts it
to all the nodes in the network for validation.
Only when the transaction is validated by all the nodes in the network,
the cost of the transaction is subtracted from the sender's wallet,
the transaction BCC are sent to the recipient's wallet,
the transaction gets its nonce,
the transaction is added into the sender's wallet transactions list,
and the transaction is ready to be added into the block.

If the validation of the transaction fails,
the transaction object is being deleted and a validation error occurs.

The validation of the transaction includes
both the verification of its signature,
as well as a check that the sender node has sufficient balance.

Cost of the transaction:
For every BCC sent, there is an additional 3% fee.
For messages, there is a cost of 1 BCC per character.